+++ src/RadixDlt.CoreApiSdk/Model/ObjectSubstateTypeReference.cs
@@ -1,0 +1,6 @@
+﻿namespace RadixDlt.CoreApiSdk.Model;
+
+public abstract partial class ObjectSubstateTypeReference
+{
+    public abstract SchemaDetails GetSchemaDetails();
+}